requirements.txt
discord.py
colorthief
requests
pytz
